analysis: fix coverage index; and generate clients, #TASK-5645

+++ opencga-analysis/src/main/java/org/opencb/opencga/analysis/alignment/AlignmentCoverageAnalysis.java
@@ -20,10 +20,14 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.hadoop.fs.FileUtil;
 import org.opencb.commons.datastore.core.ObjectMap;
+import org.opencb.commons.datastore.core.Query;
 import org.opencb.commons.datastore.core.QueryOptions;
+import org.opencb.commons.datastore.core.QueryParam;
 import org.opencb.opencga.analysis.tools.OpenCgaToolScopeStudy;
 import org.opencb.opencga.analysis.wrappers.deeptools.DeeptoolsWrapperAnalysisExecutor;
+import org.opencb.opencga.catalog.db.api.FileDBAdaptor;
 import org.opencb.opencga.catalog.exceptions.CatalogException;
+import org.opencb.opencga.core.api.ParamConstants;
 import org.opencb.opencga.core.exceptions.ToolException;
 import org.opencb.opencga.core.models.alignment.CoverageIndexParams;
 import org.opencb.opencga.core.models.common.Enums;
@@ -45,15 +49,14 @@ import static org.opencb.opencga.core.tools.OpenCgaToolExecutor.EXECUTOR_ID;
 public class AlignmentCoverageAnalysis extends OpenCgaToolScopeStudy {
 
     public final static String ID = "coverage-index-run";
-    public final static String DESCRIPTION = "Compute the coverage from a given alignment file, e.g., create a .bw file from a .bam file";
+    public final static String DESCRIPTION = "Compute the coverage from a given alignment file, e.g., create a "
+            + AlignmentConstants.BIGWIG_EXTENSION + " file from a " + AlignmentConstants.BAM_EXTENSION + " file";
 
     @ToolParams
     protected final CoverageIndexParams coverageParams = new CoverageIndexParams();
 
     private File bamCatalogFile;
     private File baiCatalogFile;
-
-    private Path bwPath;
 
     protected void check() throws Exception {
         super.check();
@@ -67,7 +70,7 @@ public class AlignmentCoverageAnalysis extends OpenCgaToolScopeStudy {
         bamCatalogFile = getFile(coverageParams.getBamFileId(), "BAM");
 
         // Check if the input file is .bam
-        if (!bamCatalogFile.getName().endsWith(".bam")) {
+        if (!bamCatalogFile.getName().endsWith(AlignmentConstants.BAM_EXTENSION)) {
             throw new ToolException("Invalid input alignment file '" + coverageParams.getBamFileId() + "' (" + bamCatalogFile.getName()
                     + "): it must be in BAM format");
         }
@@ -75,44 +78,55 @@ public class AlignmentCoverageAnalysis extends OpenCgaToolScopeStudy {
         // Getting BAI file
         if (StringUtils.isEmpty(coverageParams.getBaiFileId())) {
             // BAI file ID was not provided, looking for it
-            baiCatalogFile = getBaiFile(bamCatalogFile.getName() + ".bai");
+            baiCatalogFile = getBaiFile(bamCatalogFile.getName() + AlignmentConstants.BAI_EXTENSION);
         } else {
             // Getting the BAI file provided
             baiCatalogFile = getFile(coverageParams.getBaiFileId(), "BAI");
         }
+        logger.info("BAI file ID = {}; path = {}", baiCatalogFile.getId(), Paths.get(baiCatalogFile.getUri()));
 
         // Checking filenames
-        if (!baiCatalogFile.getName().equals(bamCatalogFile.getName() + ".bai")) {
-            throw new ToolException("Filenames mismatch, BAI file name consists of BAM file name plus the extension .bai; BAM filename = "
-                    + bamCatalogFile.getName() + ", BAI filename = " + baiCatalogFile.getName());
+        if (!baiCatalogFile.getName().equals(bamCatalogFile.getName() + AlignmentConstants.BAI_EXTENSION)) {
+            throw new ToolException("Filenames mismatch, BAI file name must consist of BAM file name plus the extension "
+                    + AlignmentConstants.BAI_EXTENSION + "; BAM filename = " + bamCatalogFile.getName() + ", BAI filename = "
+                    + baiCatalogFile.getName());
         }
 
         // Sanity check: window size
-        logger.info("{}: checking window size {}", ID, coverageParams.getWindowSize());
+        logger.info("Checking window size {}", coverageParams.getWindowSize());
         if (coverageParams.getWindowSize() <= 0) {
             coverageParams.setWindowSize(Integer.parseInt(COVERAGE_WINDOW_SIZE_DEFAULT));
-            logger.info("{}: window size is set to {}", ID, coverageParams.getWindowSize());
+            logger.info("Window size is set to {}", coverageParams.getWindowSize());
         }
-
-        // Path where the BW file will be created
-        bwPath = getOutDir().resolve(bamCatalogFile.getName() + ".bw");
     }
 
     @Override
     protected void run() throws Exception {
         setUpStorageEngineExecutor(study);
 
-        logger.info("{}: running with parameters {}", ID, coverageParams);
+        logger.info("Running with parameters {}", coverageParams);
 
         step(() -> {
 
+            // Path where the BW file will be created
+            Path bwPath = getOutDir().resolve(bamCatalogFile.getName() + AlignmentConstants.BIGWIG_EXTENSION);
+
             // In order to run "deeptools bamCoverage", both BAM and BAI files must be located in the same folder
-            // to do that symbolic links will be created
-            Files.createSymbolicLink(getOutDir().resolve(bamCatalogFile.getName()), Paths.get(bamCatalogFile.getUri()));
-            Files.createSymbolicLink(getOutDir().resolve(baiCatalogFile.getName()), Paths.get(baiCatalogFile.getUri()));
+            // Check if both BAM and BAI files are located in the same folder otherwise these files will symbolic-link temporarily
+            // in the job dir to compute the BW file; then BAM and BAI symbolic links will be deleted from the job dir
+            Path bamPath = Paths.get(bamCatalogFile.getUri()).toAbsolutePath();
+            Path baiPath = Paths.get(baiCatalogFile.getUri()).toAbsolutePath();
+            if (bamPath.getParent() != baiPath.getParent()) {
+                logger.info("BAM and BAI files must be symbolic-linked in the job dir since they are in different directories: {} and {}",
+                        bamPath, baiPath);
+                bamPath = getOutDir().resolve(bamCatalogFile.getName()).toAbsolutePath();
+                baiPath = getOutDir().resolve(baiCatalogFile.getName()).toAbsolutePath();
+                Files.createSymbolicLink(bamPath, Paths.get(bamCatalogFile.getUri()).toAbsolutePath());
+                Files.createSymbolicLink(baiPath, Paths.get(baiCatalogFile.getUri()).toAbsolutePath());
+            }
 
             Map<String, String> bamCoverageParams = new HashMap<>();
-            bamCoverageParams.put("b", Paths.get(bamCatalogFile.getUri()).toAbsolutePath().toString());
+            bamCoverageParams.put("b", bamPath.toString());
             bamCoverageParams.put("o", bwPath.toAbsolutePath().toString());
             bamCoverageParams.put("binSize", String.valueOf(coverageParams.getWindowSize()));
             bamCoverageParams.put("outFileFormat", "bigwig");
@@ -133,16 +147,20 @@ public class AlignmentCoverageAnalysis extends OpenCgaToolScopeStudy {
                         + ") was not create, please, check log files.");
             }
 
-            // Remove symbolic links
-            Files.delete(getOutDir().resolve(bamCatalogFile.getName()));
-            Files.delete(getOutDir().resolve(baiCatalogFile.getName()));
+            // Remove symbolic links if necessary
+            if (getOutDir().resolve(bamCatalogFile.getName()).toFile().exists()) {
+                Files.delete(getOutDir().resolve(bamCatalogFile.getName()));
+            }
+            if (getOutDir().resolve(baiCatalogFile.getName()).toFile().exists()) {
+                Files.delete(getOutDir().resolve(baiCatalogFile.getName()));
+            }
         });
     }
 
     private File getFile(String fileId, String msg) throws ToolException {
         OpenCGAResult<File> fileResult;
         try {
-            logger.info("{}: checking {} file ID '{}'", ID, msg, fileId);
+            logger.info("Checking {} file ID '{}'", msg, fileId);
             fileResult = catalogManager.getFileManager().get(getStudy(), fileId, QueryOptions.empty(), getToken());
         } catch (CatalogException e) {
             throw new ToolException("Error accessing " + msg + " file '" + fileId + "' of the study " + getStudy() + "'", e);
@@ -156,8 +174,9 @@ public class AlignmentCoverageAnalysis extends OpenCgaToolScopeStudy {
     private File getBaiFile(String filename) throws ToolException {
         OpenCGAResult<File> fileResult;
         try {
-            logger.info("{}: looking BAI file ID '{}'", ID, filename);
-            fileResult = catalogManager.getFileManager().get(getStudy(), filename, QueryOptions.empty(), getToken());
+            logger.info("Looking for BAI file ID from name '{}'", filename);
+            Query query = new Query(FileDBAdaptor.QueryParams.NAME.key(), filename);
+            fileResult = catalogManager.getFileManager().search(getStudy(), query, QueryOptions.empty(), getToken());
         } catch (CatalogException e) {
             throw new ToolException("Error accessing BAI file name '" + filename + "' of the study " + getStudy() + "'", e);
         }
@@ -171,7 +190,7 @@ public class AlignmentCoverageAnalysis extends OpenCgaToolScopeStudy {
                 } else {
                     // Get the most recent according to the creation date
                     // Creation date keeps the format: 20240212151427
-                    if (Integer.parseInt(file.getCreationDate()) > Integer.parseInt(selectedFile.getCreationDate())) {
+                    if (file.getCreationDate().compareTo(selectedFile.getCreationDate()) > 0) {
                         selectedFile = file;
                     }
                 }
